// client/messages.ts
export enum ClientMessage {
  Welcome = "Welcome Player ----> ",
  Divider = "------------------------",
  Intro = "You will receive a random number that you need to make divisible by 3 by adding either 1, 0, or -1. Play your cards right and you could be the winner!",
  GameNumber = "GAME NUMBER: ",
  MakeYourMove = "Make a move, remember you can only send 1, 0,  -1",
  YourMove = "The move you made is ----> ",
}